+++ src/main/java/kr/co/groupworks/entity/cis/Employee.java
@@ -50,9 +50,9 @@ public class Employee {
     private LocalDateTime joinDate;
 
     @Column(name = "salary")
-    private Integer salary;
+    private Long salary;
 
     @Column(name = "supervisor_id")
-    private Integer supervisorId;
+    private Long supervisorId;
 
 }
